Add test for pareto front sizes
